*: try to reduce the probability of trigger TestDefaultValuesAreSettable unstable
