+++ pkg/sessionctx/variable/slow_log_test.go
@@ -25,9 +25,7 @@ import (
 	"github.com/pingcap/tidb/pkg/sessionctx"
 	"github.com/pingcap/tidb/pkg/sessionctx/slowlogrule"
 	"github.com/pingcap/tidb/pkg/sessionctx/stmtctx"
-	"github.com/pingcap/tidb/pkg/sessionctx/vardef"
 	"github.com/pingcap/tidb/pkg/sessionctx/variable"
-	"github.com/pingcap/tidb/pkg/testkit"
 	"github.com/pingcap/tidb/pkg/util/execdetails"
 	"github.com/pingcap/tidb/pkg/util/mock"
 	"github.com/stretchr/testify/require"
@@ -538,59 +536,4 @@ func TestParseGlobalSlowLogRules(t *testing.T) {
 	require.Equal(t, allConditionFields, slowLogRuleSet.RulesMap[variable.UnsetConnID].Fields)
 	// Conn_ID: 789
 	require.Equal(t, uint64(789), slowLogRuleSet.RulesMap[789].Rules[0].Conditions[0].Threshold)
-}
-
-func TestSlowLogMaxPerSec(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-
-	// default value
-	tk.MustQuery(`show variables like "tidb_slow_log_max_per_sec"`).Check(
-		testkit.Rows("tidb_slow_log_max_per_sec 0"),
-	)
-	_, err := tk.Exec(`select @@SESSION.tidb_slow_log_max_per_sec`)
-	require.Equal(t, "[variable:1238]Variable 'tidb_slow_log_max_per_sec' is a GLOBAL variable", err.Error())
-	tk.MustQuery(`select @@Global.tidb_slow_log_max_per_sec`).Check(
-		testkit.Rows("0"),
-	)
-
-	// test errors
-	_, err = tk.Exec(`set session tidb_slow_log_max_per_sec="0"`)
-	require.Equal(t, "[variable:1229]Variable 'tidb_slow_log_max_per_sec' is a GLOBAL variable and should be set with SET GLOBAL", err.Error())
-	_, err = tk.Exec(`set global tidb_slow_log_max_per_sec=""`)
-	require.Equal(t, "[variable:1232]Incorrect argument type to variable 'tidb_slow_log_max_per_sec'", err.Error())
-	_, err = tk.Exec(`set global tidb_slow_log_max_per_sec="1.23"`)
-	require.Equal(t, "[variable:1232]Incorrect argument type to variable 'tidb_slow_log_max_per_sec'", err.Error())
-
-	// test warnings
-	_, err = tk.Exec(`set global tidb_slow_log_max_per_sec="-1"`)
-	tk.MustQuery("SHOW WARNINGS").Check(testkit.Rows("Warning 1292 Truncated incorrect tidb_slow_log_max_per_sec value: '-1'"))
-	tk.MustQuery(`select @@Global.tidb_slow_log_max_per_sec`).Check(
-		testkit.Rows("0"),
-	)
-	tk.MustExec(`set global tidb_slow_log_max_per_sec="1234567"`)
-	tk.MustQuery("SHOW WARNINGS").Check(testkit.Rows("Warning 1292 Truncated incorrect tidb_slow_log_max_per_sec value: '1234567'"))
-	tk.MustQuery(`show variables like "tidb_slow_log_max_per_sec"`).Check(
-		testkit.Rows("tidb_slow_log_max_per_sec 1000000"),
-	)
-
-	// normal
-	tk.MustExec(`set global tidb_slow_log_max_per_sec="2"`)
-	require.True(t, vardef.GlobalSlowLogRateLimiter.Allow())
-	require.True(t, vardef.GlobalSlowLogRateLimiter.Allow())
-	require.False(t, vardef.GlobalSlowLogRateLimiter.Allow())
-	tk.MustQuery(`show variables like "tidb_slow_log_max_per_sec"`).Check(
-		testkit.Rows("tidb_slow_log_max_per_sec 2"),
-	)
-	tk.MustQuery(`select @@Global.tidb_slow_log_max_per_sec`).Check(
-		testkit.Rows("2"),
-	)
-	// no limit
-	tk.MustExec(`set global tidb_slow_log_max_per_sec="0"`)
-	require.True(t, vardef.GlobalSlowLogRateLimiter.Allow())
-	require.True(t, vardef.GlobalSlowLogRateLimiter.Allow())
-	require.True(t, vardef.GlobalSlowLogRateLimiter.Allow())
-	tk.MustQuery(`show variables like "tidb_slow_log_max_per_sec"`).Check(
-		testkit.Rows("tidb_slow_log_max_per_sec 0"),
-	)
 }
